+++ src/Device/PowerAction/powerActionCommand.ts
@@ -1,10 +1,11 @@
 import ICommand from "../../Command/ICommand";
-import { getOrganization, getOrganizationUid, ORGANIZATION_UID_OPTION } from "../../Organization/organizationFacade";
+import { getOrganization, ORGANIZATION_UID_OPTION } from "../../Organization/organizationFacade";
 import { DEVICE_UID_OPTION, getDeviceUid, typeMap } from "../deviceFacade";
 import { CommandLineOptions } from "command-line-args";
 import { createOrganizationRestApi } from "../../helper";
 import chalk from "chalk";
 import { getActionType } from "../deviceFacade";
+import { getOrganizationUidOrDefaultOrSelect } from "../../Applet/Upload/appletUploadCommandHelper";
 
 export const powerAction: ICommand = {
 	name: 'power-action',
@@ -16,7 +17,7 @@ export const powerAction: ICommand = {
 	],
 	commands: [],
 	async run(options: CommandLineOptions) {
-		const organizationUid = await getOrganizationUid(options);
+		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 		const organization = await getOrganization(organizationUid);
 		const restApi = createOrganizationRestApi(organization);
 		const deviceUid = await getDeviceUid(restApi, options);