+++ src/Organization/Get/organizationGetCommand.ts
@@ -1,8 +1,9 @@
 import chalk from 'chalk';
 import * as Debug from 'debug';
 import { CommandLineOptions } from "command-line-args";
-import { getOrganizationUid, getOrganization, ORGANIZATION_UID_OPTION } from '../organizationFacade';
+import { getOrganization, ORGANIZATION_UID_OPTION } from '../organizationFacade';
 import ICommand from '../../Command/ICommand';
+import { getOrganizationUidOrDefaultOrSelect } from '../../Applet/Upload/appletUploadCommandHelper';
 const debug = Debug('@signageos/cli:Organization:get');
 
 export const organizationGet: ICommand = {
@@ -14,7 +15,7 @@ export const organizationGet: ICommand = {
 	commands: [],
 	async run(options: CommandLineOptions) {
 		debug('Organization get');
-		const organizationUid = await getOrganizationUid(options);
+		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 		const organization = await getOrganization(organizationUid);
 		console.log(chalk.yellow(JSON.stringify(organization, undefined, 2)));
 	},