+++ tests/unit/Applet/Upload/appletUploadCommandHelperMock.ts
@@ -1,7 +1,7 @@
 import * as sinon from 'sinon';
 
 export default {
-	getOrganizationUidAndUpdateConfig: sinon.fake(() => 'fakeOrganizationUid'),
+	getOrganizationUidOrDefaultOrSelect: sinon.fake(() => 'fakeOrganizationUid'),
 	getAppletBinaryFileAbsolutePath: sinon.fake(() => 'fakeBinaryFileAbsolutePath'),
 	getAppletDirectoryAbsolutePath: sinon.fake(() => 'fakeDirectoryAbsolutePath'),
 	getAppletEntryFileAbsolutePath: sinon.fake(() => 'fakeEntryFileAbsolutePath'),